requirements.txt
# Frontend Requirements for Medical Chatbot
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0

# Backend Requirements for Medical Chatbot
# LangChain Core
langchain
langchain-core
langchain-community
langgraph>=0.2.0

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Document Processing
pypdf
pymupdf
tqdm

# Vector Stores
faiss-cpu
chromadb
sentence-transformers

# Additional Libraries
numpy
scikit-learn

# FastAPI Server
fastapi
uvicorn
pydantic